daily/20201209/example_googleapi/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

metashape
httpx
